+++ Demo/Sources/Demos/Demo Screens/CollectionViewAppearance.swift
@@ -33,13 +33,15 @@ extension LayoutDescription
             )
             
             $0.layout = .init(
+                headerToFirstSectionSpacing: 20.0,
                 interSectionSpacingWithNoFooter: 20.0,
                 interSectionSpacingWithFooter: 20.0,
                 sectionHeaderBottomSpacing: 15.0,
                 itemSpacing: 10.0,
                 itemToSectionFooterSpacing: 10.0
             )
-            
+
+            $0.stickyListHeader = true
             $0.stickySectionHeaders = true
             
             configure(&$0)